pyproject.toml
[tool.poetry]
name = "neurorat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
torch = "^2.3.0"
torchvision = "^0.18.0" 
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
tensorboard = "^2.16.2"
torch-tb-profiler = "^0.4.3"
gymnasium = "0.28.1"
shapely = "^2.0.4"
pandas = "^2.2.2"
svgpath2mpl = "^1.0.0"
pygame = "^2.5.2"
matplotlib = "3.8.0"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.2.1"
stable-baselines3 = "2.0.0"
numpy = "^1.26.4"
numba = "^0.60.0"
scikit-learn = "^1.5.1"
deap = "^1.4.1"
shiny = "^1.0.0"
shinywidgets = "^0.3.2"
plotly = "^5.23.0"
faicons = "^0.2.2"
nbformat = "^5.10.4"
trimesh = "^4.4.7"
chardet = "^5.2.0"
pygltflib = "^1.16.2"
pythreejs = "^2.4.2"
pyrender = "^0.1.45"
cairosvg = "^2.7.1"
geopandas = "^1.0.1"
pyglet = "<2"
pyarrow = "^18.0.0"
fastparquet = "^2024.11.0"
torch-geometric = "^2.6.1"
seaborn = "^0.13.2"
xgboost = "^3.0.0"
pynvml = "^12.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
